Only warn if a _different_ ephemeral device is attached
When we retry a task, we run the Run method again.  But in this case,
the run method actually populated some default values.  Only warn if the
values we are populating are different, to avoid spurious warnings.

We probably need a stronger lifecycle - for example having a Validate
method would probably be helpful.

Fix #48