CI: Add container build
Add a simple Dockerfile that can build kops, and a short README on how
to build it. This generates a 425M image, which is about as tight as I
could get it in a short amount of work without just busting apart the
golang:1.6-alpine image to compress the Go layer itself. It's good
enough for what I'm planning to use it for: CI builds.

Along the way: Fix-up .gitignore to include more from k8s/.gitignore
(include editor ones, and terraform output dir).