Create stable channel for kops
The idea being we don't necessarily always want to pick up the latest
change immediately, without validating it first.

Also this is an easy way to recommend the latest AMI without having to
search for it.  We also create an alias to the owner id, so that it is
readable / verifiable.