Fix bug in security group matching
We were matching only by the name tag, instead of using the group-name

If users had an SG with the same Name in different VPCs we could have
matched SGs in both VPCs.  This seems unlikely, and particularly tricky
to then get through the remaining sanity checks (even in the find
function itself).

Fix #813