Fix a DHCP options leak in the case where there's no default VPC
If you create a VPC in a region where you have no "default-looking"
DHCP option set (e.g. a region with no default VPC), kops will leak
the "default-looking" VPC when it reassociates the tagged one.

The solution is to carefully try to determine if replaced DHCP option
set is default-looking (no tags, defaults set) and courtesy-delete it
if it is, ignoring dep violations.