Don't fail elastic IP task if TagOnSubnet is not set
We don't yet have a strong strategy for how to handle this if we are
interrupted, but the same problem exists with taggging in general.
Likely we'll need to write something like a WAL-log, and then check for
interrupted in-flight objects next time we run.

In any case, we shouldn't treat not using TagOnSubnet as an error.