Expose kube-apiserver audit log to host volume
This commit exposes kube-apiserver's audit log to the host as a host
mapping.

PR #1872 gave the ability to users to define a custom log path for the
apiserver to write its audit logs to. Prior to this commit, the log file
would stay within the container's filesystem, and getting access to it from
outside the container was a nuisance.

This change allows a logging aggregator, like fluentd, to be able
to read and tail this log from outside the kube-apiserver container.