Bump canal addon version
In PR [#2635](https://github.com/kubernetes/kops/pull/2635) the version
of calico was updated as part of canal.  What this PR missed was bumping
the version of the addon.  This meant that if you upgraded a kops
cluster from v1.6.0 to v1.6.1 the new version of calico was not picked
up because `channels apply` does not think the daemonset needs to be
updated.