FIX #4055 dns controller issue with gossip kops1.8
Revert part of commit f157ccc to fix issue #4055. The dns-controller
switch statement to set the cloud providor is enclosed in an if
statement. It will never be set to anything other than gossip if the
gossip protocol is used.  I believe this is not the intended behavior
as using the gossip protocol with any of the cloud providers in the
switch is still valid configuration.