Adding additional tags to various AWS components
This PR adds the base tags to DHCP Options, IGW, and Route Tables.
These components are not tagged correctly, and this can cause issues
with deletion. Name tags are not added to shared resources, as we allow
shared resources to have maintained names.  A owned/shared tags with the
syntax  "kubernetes.io/cluster/$CLUSTERNAME" = "owned" is added to the
resources as well.  We are maintaining the Name tag value for private
route tables, as these resources do not use the standard value.