attempt to use the private dns name first if it resolves
rolling update will fail the first time you try to upgrade an existing cluster using private dns. The masters will be using old certs that haven't had the ELB name added to them. To avoid such failure, we can check to see if the private dns does resolve correctly where we're running kops from and if so, use that.