Don't use ELB DNS name for internal ELBs
We introduced some new heuristics in #3941 that meant we would access
private DNS names via the ELB's name, just as we do with gossip names.

But it's also possible to set up a VPN or DirectConnect and access a
private HostedZone directly, and in this case we don't want to use the
ELB DNS Name.  We recognize this case because the ELB can be set to
Internal (and probably should be, for minimal attack surface!)

We'll probably have to introduce a field for this, but hopefully this
heuristic is sufficient to unblock the release.