Use docker ps -n 1 to check docker health
Under heavy load, `docker ps` takes quite a while to return. That means that you may end up restarting docker daemon and killing a bunch of pods (I've been there)
Using `docker ps -n 1` allow to check that the daemon is running, but it is more lightweight and it should take less time.