Handle storage flag on COS more carefully
Only clear the flag if there is a docker config file, so that we can
continue to set the storage flag on older COS images.  We could be
smarter about checking if the storage driver is actually set in the
docker config, but for now we just start by logging it.