nodeup: Use https for docker download
Both domains support https, so we can just use it.

I've tried to download all files via https and it works fine. I used,
with this commit as HEAD, this simple script:

	for i in $( git show | grep https | awk ' { print $3 }' | tr -d '",' | grep ^https:\/\/ )
	do
		echo $i; timeout 3 wget  $i
	done

Then, just compared the number of downloaded files and URLs used in the
for loop (there is one file, docker-ce_17.09.0~ce-0~debian_amd64.deb.1,
created with a ".1" instead as ".deb" because the distribution is not in
the filename for those).

So, besides this should work, I've tried for all those files and it
works :)