Introduce bundles
A bundle is a set of manifests (including versions) for the components
of a cluster.  This means we can update e.g. calico without having to
release a new version of kops.

The bundle spec itself is being developed at
https://github.com/GoogleCloudPlatform/k8s-cluster-bundle

We introduce a new field into the cluster.spec, "bundle" - which
specifies a bundle version, which in turn determines the versions of
all the addons in use.  This means that the operator still has control
over which version is in use.

We extend `kops cluster create` and `kops cluster upgrade` to
automatically update the version, driven by the channel.