Fixes spurious LoadBalancer change when using ACM Certificate
When using `.spec.api.loadBalancer.sslCertificate`, subsequent `kops update cluster` report the following spurious change:
```
Will modify resources:
  LoadBalancer/api.ho-prod-a.us-east-1.k8s.tune.com
  	Listeners           	 {443: {"InstancePort":443,"SSLCertificateID":""}} -> {443: {"InstancePort":443,"SSLCertificateID":"arn:aws:acm:us-east-1:000000000000:certificate/..."}}
```

This is because we werent keeping track of the ACM Certificate when describing the "actual" load balancer status.

This fixes that behavior, and now kops accurately reports no changes to be made.