Service Address Check
The current implementation assumes you have specified a service address within your config rather than using being filled in by the defaults. This PR only checks when service address has been specified. Alternatively if people feel that this should be 'enforce' at the very least we should check the reference is not nil so we dont throw a panic.