Add suggested alias for bazelrc import location
See https://github.com/bazelbuild/bazel/issues/6319

The default bazelrc location is changing.  So that we don't break
during the transition, we set up an alias so that bazel versions that
load from the new location are redirected to the old location.

Also add a TODO to move to the new location once bazel 0.18 is
prevalent.