Fail fast if io2 iops to size ratio is too high
In AWS the ratio between volume IOPS and volume size must be at most 50,
otherwise volume will fail creating. See
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/EBSVolumeTypes.html,
specifically "_The maximum ratio of provisioned IOPS to requested volume size
(in GiB) is 50:1. For example, a 100 GiB volume can be provisioned with up to
5,000 IOPS._"

This commit adds the option of failing fast when creating a new cluster if the
ratio is higher than 50. Previously kops would send the API request to AWS, fail
and repeat until the timeout was reached.