Set metrics server to use insecure tls
In 0.3.x, the secure kubelet port with auth is enabled by default. Use
of the insecure port is deprecated and may even be removed as soon as
the next release. See:
* https://github.com/kubernetes-incubator/metrics-server/releases/tag/v0.3.0
* https://github.com/kubernetes-incubator/metrics-server/releases/tag/v0.3.1

That said, metrics-server now uses webhook authentication so kubelet
would need webhook authentication enabled (which we don’t). This flag
enables, that serviceaccount tokens to be used to authenticate against
the kubelet: See:
* https://kubernetes.io/docs/reference/access-authn-authz/webhook/
* https://github.com/kubernetes/kops/pull/5508
* https://github.com/kubernetes/kops/issues/5176#issuecomment-402454053
* https://github.com/kubernetes-incubator/metrics-server/issues/175

Currently, even if we were to use certificates, this provides a
challenge since metrics server will generate it’s own self-signed
certificates since we don’t set `--tls-cert-file` and
`--tls-private-key-file`. Related:
* https://github.com/kubernetes-incubator/metrics-server/issues/25
* https://github.com/kubernetes-incubator/metrics-server/issues/146