Kubelet API RBAC Manifest
I was mistaken how the key in the namespace is used, looks like you can only have the one key per addon as the anonnotations on the namespace are addon_name -> version, so if you already have a old rbac.addons.k8s.io referenced it won't work. I didn't notice before as the cluster was new.