Fix docker-healthcheck to work around Docker bug.
This is a workaround to better detect moby/moby#38642 when Docker starts
up and remains stuck.  In this case `docker ps` will return nothing and
exit 0, but no container can actually start.  A better (but more expensive
and more intrusive test) would be to `docker run --rm` some cheap test
container to confirm we can actually start a container.