Remove docker-prestart hook
This breaks networking if Docker is restarted
(https://github.com/kubernetes/kops/issues/6191).

The Docker issue linked in the hook's comments has been closed for
over 3 years, and this workaround has not been used by upstream
Kubernetes for over a year:

  https://github.com/kubernetes/kubernetes/commit/5f9735de53c0a7f1334d91b59bcd87c6fbe2a06f#diff-af1d281c3ce49f7bfe110f7c64c96fdc

It therefore seems unlikely that this hook is still necessary.