Sync data-types for webhook config with upstream
This is going to be in componentconfig soon, so it would be nice to
have the same types.

These flags were recently mapped in #6361 and have not yet been in a
release - it's now or never!  (Though technically it is only the
AuditWebhookBatchThrottleEnable that won't parse identically)

Also added tests!