Support custom kube-scheduler configuration
Mentioned in https://github.com/kubernetes/kops/issues/6942

This change allows using the --config flag to set additional options,
theese are marked as depecrated as flags
(https://kubernetes.io/docs/reference/command-line-tools-reference/kube-scheduler/)
so I thought that it might be better to have them in a config file.

It also makes it easy to add more.