GCE tutorial markdown formatting
Hi! First of all I'd like to thank you for making this! I'm eager to try it out o/

The current `master` version of the document contains a few formatting issues that makes it hard to read. The proposed changes attempt to make formatting better to read and concise. Rendering the Github markdown will provide better "before and after" view of the changes.