"Force" k8s 1.11.10 in stable channel
Bringing this from the alpha channel (#7423) to the stable channel.

As discussed in office hours, and from the notes on #7423:

Even 1.11 and 1.12 are technically unsupported, but this should get
users off the very oldest unsupported versions of k8s.

Users that want to stay on those versions can pass the
KOPS_RUN_OBSOLETE_VERSION env var.