Use /opt/cni/bin on all distros
We used to remap the cni-bin to /home/kubernetes/bin on COS, but that
then requires us to change the CNI manifests also to write to the new
location.

Instead we can use /opt/cni/bin on all distros, now that we are making
it writeable everywhere with a bind mount.