Fix verify-staticcheck prow job
The new job was missing the generated files from go-bindata.

I noticed some of the prow jobs use make and others call hack/ scripts directly.

I think it'd be better to use make so we can more easily add dependencies like this, therefor I'd like to propose that we merge this, update the jobs to call make targets, then remove the make command I added to hack/verify-staticcheck.sh

By fixing the existing job as well as the make target, we can unblock other PRs and not get another swarm of failure notifications when the job gets updated to use make.