Update CentOS 7 AMI
Our CentOS 7 e2e job has been failing for a long time: https://testgrid.k8s.io/sig-cluster-lifecycle-kops#kops-aws-centos-7

due to the AMI's lack of ENA/NVMe support and our tests defaulting to m5.large for the masters.
I updated the AMI in the e2e job and it appears to be successfully launching the cluster (still experiencing aws volume flakes).
This updates our docs to match the new AMI ID. While the centos website is not up to date, this AMI ID can be found with the documented aws cli query.