Move cloudbuild.yaml to repo root
This file is used by the push-to-staging postsubmit job that builds images and pushes them to a staging repo.
When it was originally added, my theory for why the job was failing [0] is that we were only uploading the ci/postsubmit/push-to-staging directory to GCS for use in cloudbuild, when in reality we need to push the entire kops repo so it can run `make kops-controller-push`.

A while back I had updated the job to use the repo root but hadnt created the cloudbuild.yaml file there yet, so instead we've been getting a cloudbuild.yaml not found error [1].

This moves the cloudbuild.yaml into the repo root so that the current job configuration will recognize it and hopefully be able to run `make kops-controller-push` in GCB.

[0] https://prow.k8s.io/view/gcs/kubernetes-jenkins/logs/kops-postsubmit-push-to-staging/1209130004504383489
[1] https://prow.k8s.io/view/gcs/kubernetes-jenkins/logs/kops-postsubmit-push-to-staging/1213230826582446081