Remove .git dependency from get_workspace_status.sh
The image pushing postsubmit job [0] is one step closer to succeeding.

It currently fails because when the builder image uploads the kops directory to GCS for to be consumed by GCB, it excludes the .git directory [1].

This causes the job to fail because `make kops-controller-push` uses bazel which runs get_workspace_status.sh which aborts if the git commands fail.

The prow job doesnt contain much output but the GCB logs can be seen in GCS [2].

This PR removes the dependency on git commands.
I dont think any of the logic in the script will change because we already set VERSION [3] which is the only variable in get_workspace_status.sh that depends on git commands and is used to build and push the kops controller image.
[0] https://testgrid.k8s.io/sig-cluster-lifecycle-kops#kops-postsubmit-push-to-staging
[1] https://github.com/kubernetes/test-infra/blob/18391d89866518581e4b6de2546e16ba2b08a3be/images/builder/main.go#L87
[2] gs://k8s-staging-kops-gcb/logs/log-a7dc3a24-97cd-42fe-bec3-971dc78a0e3a.txt
[3] https://github.com/kubernetes/kops/blob/b1276ac8353279c2d1de7beea4e6ec3713d83f6c/cloudbuild.yaml#L10