Remove hack/.packages
This was originally used to speed up `go list` when passing packages to `go test` but was removed when `go test` began ignoring the vendor directory by default.

This was the only reference I could find in the repo to hack/.packages so I think it is safe to remove.
We'll need to remove the prow job as well.